google_fonts {showtextdb}	R Documentation

Meta-information for the Google Fonts

Description

This function provides information of Google Fonts that can be used in the font_install() function. It will try to include all font faces available for a specified family name.

Usage

google_fonts(name, ...)

Arguments

name
Name of the font that will be searched in Google Fonts.

...
Other parameters passed to curl::curl_fetch_memory().

Author(s)

Yixuan Qiu <https://statr.me/>

Examples

opensans = google_fonts("Open Sans")
print(opensans)

## Not run: 
## Install the Open Sans font to the showtexdb package
font_install(opensans)

## End(Not run)

[Package showtextdb version 3.0 Index]