source_han {showtextdb}	R Documentation

Meta-information for the Source Han Sans/Serif Fonts

Description

These functions provide information of the Source Han Sans/Serif fonts that can be used in the font_install() function. Source Han Sans/Serif fonts provide complete support for the CJK (Chinese, Japanese, and Korean) characters.

Usage

source_han_sans(lang = c("CN", "TW", "JP", "KR"))

source_han_serif(lang = c("CN", "TW", "JP", "KR"))

Arguments

lang
Language of the font. "CN" for simplified Chinese, "TW" for traditional Chinese, "JP" for Japanese, and "KR" for Korean.

Author(s)

Yixuan Qiu <https://statr.me/>

Examples

## Not run: 
## Install Source Han Sans font (by default Simplified Chinese)
## to the showtexdb package
font_install(source_han_sans())

## Source Han Serif Japanese
font_install(source_han_serif("JP"))

## End(Not run)

[Package showtextdb version 3.0 Index]